rename: preference 위젯 폴더 이동
